chore(rxForm): refactor rxForm
**rxFormOptionTable split from rxForm as rxOptionTable**

* Stand-in replacement for `rxFormOptionTable`
* `rxFormOptionTable` has been aliased to point to `rxOptionTable`
  with console.warn message to notify developers that they need to
  update their code
  * console.warn tested in spec
* API has NOT changed in the move
  * Internal functions may have been updated/renamed, but these should
    not affect developer consumption.
* Attributes are now documented in the JSDocs rather than the README
  * provides an easier way for framework develoeprs to keep info updated
    as the JSDocs are generated from javascript comments right next to
    the code.

-----

**Moved to rxMisc**

* `rxForm.currencyToPennies`
  * `rxForm.currencyToPennies` deprecated
    * currently aliased to `rxMisc.currencyToPennies`
* `rxForm.slowClick`
  * `rxForm.slowClick` deprecated
    * currently aliased to `rxMisc.slowClick`

**Refactor rxForm Page Object**

This is to aid in the upcoming rework to rxForm and to make use of
existing page objects in other components (rxCheckbox, rxRadio, etc.)

**Deprecated Properties**

All of the following have been deprecated with notes in the test
documentation.

* `rxOptionFormTable`
  * currently aliased to `rxOptionTable`
* `rxForm.checkbox`
  * currently aliased to `htmlCheckbox`
  * `htmlCheckbox.unselect` added for backward compatibility
    * deprecated with note in test docs to use `deselect()`
    * currently aliased to `htmlCheckbox.deselect`
* `rxForm.radioButton`
  * currently aliased to `htmlRadio`
* `rxForm.dropdown`
  * currently aliased to `htmlSelect`

-----

**Updated rxForm Demo Page**

* rxForm demo updated to use `rxOptionTable`
* `rxFormOptionTable` has been documented as deprecated in `rxForm/README.md`
