+++ src/rxForm/README.md
@@ -32,18 +32,11 @@ Creates a field row wrapped in a fieldset. Used for fields which include multipl
 
 ## rxFormOptionTable
 
-Given a data object and an additional optional object for column labels, rxFormOptionTable creates a series of radio or checkbox buttons.
+[![deprecated](http://badges.github.io/stability-badges/dist/deprecated.svg)](http://github.com/badges/stability-badges)
 
-### Attributes
+**DEPRECATED - This directive has been marked as deprecated and *will be removed* in a future release of the EncoreUI framework.**
 
-- `type`: Values `radio` or `checkbox` are required.
-- `required`: For checkboxes, a `true` value means there must be at least one checkbox checked.
-- `columns`: A data object listing column copy, include labels and keys. Expressions are allowed; see the samples.
-- `data`: A data object to prefill the radio/checkbox collection. For checkboxes, checked values default to true unless `value` and `falseValue` attributes are given. See the samples.
-- `model`: The AngularJS model to tie all radios/checkboxes together.
-- `field-id`: The ID of the elements.
-- `empty-message`: (string) A default message if the data attribute is empty.
-- `disable-fn`: (Function) Optional callback function to determine if option should be disabled. Parameters `tableId`, `fieldId`, and `rowId` will be passed to the function.
+Please use [`rxOptionTable`](/#/component/rxOptionTable) as a stand-in replacement.
 
 # Services
