+++ src/rxForm/rxForm.less
@@ -23,7 +23,7 @@
 
 .rx-form-fieldset > fieldset {
     width: 100%;
-    .full-width .option-table {
+    .full-width .rx-option-table {
         width: 100%;
     }
 }
@@ -209,68 +209,6 @@
     width: 50%;
     select {
         .border-radius(0px);
-    }
-}
-
-.option-table {
-    // we want the table to not look ugly when stretched across the screen
-    width: auto;
-    // we also want a semi-consistent width, but not too wide for modals
-    min-width: 39em;
-
-    td:not(.empty-data) {
-        // remove default padding from td's so that we can make the label take up the full room
-        padding: 0;
-    }
-    input[disabled] {
-        cursor: not-allowed;
-    }
-    .disabled {
-        label {
-            cursor: default;
-        }
-        span {
-            color: @inputColorDisabled;
-        }
-    }
-    .fillWrapper {
-      box-sizing: border-box;
-      display: inline-block;
-      height: 100%;
-      width: 100%;
-      position: relative;
-
-      label {
-        box-sizing: border-box;
-        width: 100%;
-        height: 100%;
-        padding: @tableCellPadding;
-        cursor: pointer;
-
-        .flexbox();
-        // vertically centered content in label
-        .align-items(center);
-
-        .alignWrapper {
-          .flex(1 1 auto);
-        }
-      }
-    }
-    .option-table-input {
-        text-align: center;
-        line-height: 0;
-    }
-    .current td {
-        color: #3AB661;
-    }
-    // @note KL: I'd like to stick with just adding the background to the <tr>, but
-    // we need to make the selector more specific to override the table-stripe styles.
-    .selected td {
-        background: @tableRowSelected;
-        border-color: #bad4bb;
-        // this 'double' makes it so that the border-color applies to the top border as well
-        // @see http://stackoverflow.com/questions/7942212/css-table-row-border-color-with-border-collapse
-        border-style: double;
     }
 }
