+++ src/rxOptionTable/README.md
@@ -1,0 +1,7 @@
+[![unstable](http://badges.github.io/stability-badges/dist/unstable.svg)](http://github.com/badges/stability-badges)
+
+# rxOptionTable
+
+Given a data object and an additional optional object for column labels, rxFormOptionTable creates a series of radio or checkbox buttons.
+
+* <a href="/ngdocs/index.html#/api/rxOptionTable.directive:rxOptionTable" target="_blank">View API Documentation</a>
