+++ src/rxOptionTable/docs/rxOptionTable.html
@@ -1,0 +1,46 @@
+<div ng-controller="rxOptionTableCtrl">
+  <h3 class="title">Radio Option Table</h3>
+  <rx-option-table
+      id="radioOptionTable"
+      data="optionTableData"
+      columns="optionTableColumns"
+      type="radio"
+      model="radioValue"
+      field-id="optionTable"
+      selected="0"
+      class="full-width"
+      disable-fn="disableOption(tableId, fieldId, rowId)"
+      ></rx-option-table>
+
+  <p>Bound Value: {{radioValue}}</p>
+
+  <br />
+
+  <h3 class="title">Checkbox Option Table</h3>
+  <rx-option-table
+      id="checkboxOptionTable"
+      columns="optionTableColumns"
+      type="checkbox"
+      model="checkboxValues"
+      field-id="optionCheckboxTable"
+      data="optionTableCheckboxData"
+      required="true"
+      ></rx-option-table>
+
+  <p ng-repeat="val in checkboxValues">
+    Item {{$index + 1}} Value: {{val}}
+  </p>
+
+  <br />
+
+  <h3 class="title">Empty Option Table</h3>
+  <rx-option-table
+      id="emptyOptionTable"
+      columns="optionTableColumns"
+      type="checkbox"
+      model="emptyValue"
+      field-id="optionCheckboxTable"
+      data="optionTableEmptyData"
+      empty-message="You don't have any data!"
+      ></rx-option-table>
+</div>
