+++ src/rxOptionTable/docs/rxOptionTable.js
@@ -1,0 +1,74 @@
+/*jshint unused:false*/
+angular.module('demoApp')
+.controller('rxOptionTableCtrl', function ($scope) {
+    $scope.radioValue = 0;
+    $scope.checkboxValues = [true, 'unchecked'];
+
+    $scope.optionTableData = [
+        {
+            'id': 'option1',
+            'name': 'Option #1',
+            'value': 0,
+            'obj': {
+                'name': 'Nested Name 1'
+            }
+        }, {
+            'id': 'option2',
+            'name': 'Option #2',
+            'value': 1,
+            'obj': {
+                'name': 'Nested Name 2'
+            }
+        }, {
+            'id': 'option3',
+            'name': 'Option #3',
+            'value': 2,
+            'obj': {
+                'name': 'Nested Name 3'
+            }
+        }, {
+            'id': 'option4',
+            'name': 'Option #4',
+            'value': 3,
+            'obj': {
+                'name': 'Nested Name 4'
+            }
+        }
+    ];
+
+    $scope.optionTableColumns = [
+        {
+            'label': 'Name',
+            'key': 'name',
+            'selectedLabel': '(Already saved data)'
+        }, {
+            'label': 'Static Content',
+            'key': 'Some <strong>Text &</strong> HTML'
+        }, {
+            'label': 'Expression 2',
+            'key': '{{ value * 100 | number:2 }}'
+        }, {
+            'label': 'Expression 3',
+            'key': '{{ obj.name | uppercase }}'
+        }, {
+            'label': 'Expression 4',
+            'key': '{{ value | currency }}'
+        }
+    ];
+
+    $scope.optionTableCheckboxData = [
+        {
+            'name': 'Item 1'
+        }, {
+            'name': 'Item 2',
+            'value': 'checked',
+            'falseValue': 'unchecked'
+        }
+    ];
+
+    $scope.optionTableEmptyData = [];
+
+    $scope.disableOption = function (tableId, fieldId, rowId) {
+        return rowId === 'option4';
+    };
+});
