+++ src/rxOptionTable/docs/rxOptionTable.midway.js
@@ -1,0 +1,122 @@
+var _ = require('lodash');
+var rxOptionTablePage = require('../rxOptionTable.page').rxOptionTable;
+var rxMisc = require('../../rxMisc/rxMisc.page').rxMisc;
+var exercise = require('../rxOptionTable.exercise');
+
+describe('rxOptionTable', function () {
+    var component;
+
+    before(function () {
+        demoPage.go('#/component/rxOptionTable');
+    });
+
+    describe('exercise', function () {
+        describe('radio table', exercise.rxOptionTable({
+            cssSelector: "#radioOptionTable",
+            visible: true,
+            empty: false
+        }));
+
+        describe('checkbox table', exercise.rxOptionTable({
+            cssSelector: "#checkboxOptionTable",
+            visible: true,
+            empty: false
+        }));
+
+        describe('empty table', exercise.rxOptionTable({
+            cssSelector: "#emptyOptionTable",
+            visible: true,
+            empty: true
+        }));
+    });
+
+    describe('with radios', function () {
+        var columnNames = ['Name', 'Static Content', 'Expression 2', 'Expression 3', 'Expression 4'];
+
+        before(function () {
+            component = rxOptionTablePage.initialize($('#radioOptionTable'));
+        });
+
+        it('should have all column names', function () {
+            expect(component.columnNames).to.eventually.eql(columnNames);
+        });
+
+        it('should return column data as text by default', function () {
+            var textData = ['$0.00', '$1.00', '$2.00', '$3.00'];
+            expect(component.columnData('Expression 4')).to.eventually.eql(textData);
+        });
+
+        it('should return column data as defined by a custom function', function () {
+            var penniesData = [0, 100, 200, 300];
+            var penniesFn = function (cellElements) {
+                return cellElements.map(function (cellElement) {
+                    return cellElement.getText().then(rxMisc.currencyToPennies);
+                });
+            };
+
+            expect(component.columnData('Expression 4', penniesFn)).to.eventually.eql(penniesData);
+        });
+
+        it('should select a column by text', function () {
+            component.selectByColumnText('Expression 4', '$2.00');
+            expect(component.selections).to.eventually.eql([2]);
+        });
+
+        it('should have the current saved data in the first row', function () {
+            expect(component.row(0).isCurrent()).to.eventually.be.true;
+        });
+
+        it('should have disabled row 4', function () {
+            expect(component.disabledOptions).to.eventually.eql([3]);
+        });
+
+        describe('selected row', function () {
+            var row;
+            var cellData = ['Option #3', 'Some Text & HTML', '200.00', 'NESTED NAME 3', '$2.00'];
+            var rowData = _.zipObject(columnNames, cellData);
+
+            before(function () {
+                row = component.selectedRow;
+            });
+
+            it('should be selected', function () {
+                expect(row.isSelected()).to.eventually.be.true;
+            });
+
+            it('should not be the current saved data', function () {
+                expect(row.isCurrent()).to.eventually.be.false;
+            });
+
+            _.forEach(rowData, function (data, column) {
+                it('should have the correct data in the ' + column + ' column', function () {
+                    expect(row.cell(column)).to.eventually.equal(data);
+                });
+            });
+        });
+    });//radio
+
+    describe('with checkboxes', function () {
+        before(function () {
+            component = rxOptionTablePage.initialize($('#checkboxOptionTable'));
+        });
+
+        it('should have one row selected', function () {
+            expect(component.selections).to.eventually.eql([0]);
+        });
+
+        it('should unselect all rows', function () {
+            component.unselectAll();
+            expect(component.selections).to.eventually.be.empty;
+        });
+
+        it('should select many rows', function () {
+            component.selectMany([{ Name: 'Item 1' }, { Name: 'Item 2' }]);
+            expect(component.selections).to.eventually.eql([0, 1]);
+        });
+
+        it('should unselect many rows', function () {
+            component.unselectMany([{ Name: 'Item 1' }, { Name: 'Item 2' }]);
+            expect(component.selections).to.eventually.be.empty;
+        });
+    });
+});
